Update to 1.1.1
Includes some changes to the fix we did for windows.

Also fixed when uninstalling the path variable for php/* was not removed
now it is. Also if you did not install php the path variable is no longer
added plus if path variable for php only matches your arch so if you
install php x64 then that is the path it will go for.

Also remove unneded code which was why the path was not removed correctly.

Change-Id: Ie6302042729460955e4f00e7bf4c255178e301f8