Update installer to 1.1.2
This update includes a fix for php.ini whereby it would not look for the
ext directory in the current install path instead it was loook at the
default path c:\php

This also includes VC14 redistributable from micrsoft. It is now installed
when you install php. If you have it already installed it wont overwrite
it.

Change-Id: Idd80f7b547f3b535017fa281110fdc3c8df3044a